+++ src/View/OperationsLogic.java
@@ -2,7 +2,7 @@ package View;
 
 public interface OperationsLogic {
 
-    public double operation(int number1, int number2);
+    public double operation(double number1, double number2);
     public String getOperation();
 
 }
